requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/base.in
#
asgiref==3.2.10           # via django
django-cors-headers==3.5.0  # via -r requirements/base.in
django-environ==0.4.5     # via -r requirements/base.in
django-filter==2.4.0      # via -r requirements/base.in
django==3.0.10            # via -r requirements/base.in, django-cors-headers, django-filter, djangorestframework
djangorestframework==3.12.1  # via -r requirements/base.in
pillow==7.2.0             # via -r requirements/base.in
psycopg2-binary==2.8.6    # via -r requirements/base.in
pytz==2020.1              # via django
sqlparse==0.3.1           # via django
whitenoise==5.2.0         # via -r requirements/base.in
